+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `job-service-run` now supports `volume` for mounting host directories and named volumes (closes [#527](https://github.com/netresearch/ofelia/issues/527))
+
 ## [0.21.4] - 2026-03-17
 
 ### Fixed